Don't fail when syntax entry is missing for closing parens
In the case of some buffers, e.g. `*urlparse-temp*`, there is no syntax entry for closing parens. When trying to enable `autopair-global-mode` with such buffers open, an error results due to autopair assuming there is such an entry. This commit prevents that error by binding 'autopair-skip-close-maybe only if such a syntax entry exists.

(This behavior was observed in Emacs 24.)