Closes #3: in `autopair-insert-or-skip-quote`, handle errors without position info.
Not all errors produced `up-list` have the needed `fourth` position arg,
especially when `forward-sexp-function` is used, as is the case of
`nxml-forward-balanced-item`. In those cases, the best we can do is ignore it.

There might be a lot more situations like these, since autopair sometimes
heavily relies on the position info bundled with syntactic movement errors.